lambda/generateDocuments/requirements.txt
google-generativeai
pinecone-client